Add my changes to README
I had to erase Tan Khoa Tran's login changes, since I was already
working on a login system that worked in a different way.  Therefore, I
removed it from this file.  I apologize, but we had already agreed that
this was my job and I had already started - check the Git logs!